requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
binance_futures-1.1.0-py3-none-any.whl
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dateparser==1.1.0
Flask==2.0.2
frozenlist==1.2.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
line-notify==0.1.4
MarkupSafe==2.0.1
multidict==5.2.0
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==2.1
ujson==5.0.0
urllib3==1.26.7
websocket-client==1.3.2
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.2
